Update map to use `PathBuf` instead of `String` for filepaths
Update the map request path to use `PathBuf` instead of `String` for filepaths. This fixes the issue with the integration tests failing when being run in parallel due to the filepath being mapped being an absolute filepath, and the `PathBuf::join` call replaces the output root when called with an absolute filepath. Instead of creating a path that included the output root for the worker and the filename the ouput filepath was the absolute filepath of the mapped file plus a partition identifier. This caused issues when integration tests were being executed concurrently because both tests would be modify and reading from the same filepath, resulting in inconsistencies across executions.